+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Reconcile `v1alpha3` CR's.
+
 ## [1.4.0] - 2021-07-09
 
 ### Added